Add `stack-nix` to `stackProject` (#293)
Like `plan-nix` for `cabalProject` this gives us a way to inspect the
nix files generated by `stack-to-nix`.  It may also adds `stackProject'`
that like `cabalProject'` allows us to get `stack-nix` independently of
hsPkgs (to avoid unwanted evaluation).  This can be useful for caching
and preventing GC of evaluation time dependencies of the project.