Allow .so dynamic libraries on Darwin (#534)
* Allow .so dynamic libraries on Darwin

This code is a hack that applied only to Darwin, it collects all
required dynamic libraries into a single directory. It does this by
listing all library dirs and then linkin all dynamic libraries in them.

Before this change, it would link only `*.dylib` files. This was a
problem, because (weirdly) a lot of libraries in Nixpkgs actually
produce .so even on Darwin (and, weirdly, this somehow works).

So we now link `.so` files as well.

* Fix linking of dynamic libs with static on Darwin

This piece of code `ln`s all libraries we might want to link with into a
single directory and then hides all previous directories from Cabal.
The problem is that this code only `ln`s dynamic libraries, however we
might want to link with some library that only comes in a static
version.

So, `ln` static `.a` libraries too so that they are found when linking.