Profiled iserv for profiled builds (#574)
* Profiled iserv for profiled builds

When trying to build profiled builds, we might end up in a situation
where the library we load requires symbols from the profiled rts. In
those cases we rely on the profiled RTS linked against iserv.

Also ghc has the slightly annoying preference of adding the `-prof`
suffix to the provided iserv program, thus we need two iservs now.

* Update mingw_w64.nix

Only prof, not debug.