Add support for "latest" as a version (#934)
When using `shellFor { tools = { x = "1.2.3"; }; }` it can be a pain
to have to check if you have the latest version of `x` or if a new
one has been introduced.  This change adds support for passing
`"latest"` as the version number to anything built on top of the
`haskell-nix.hackage-project` function.  When `"latest"` is used
the newest version in the hackage.nix snapshot will be selected.