Include FSF-libre when determining free-ness (#1264)
* SPDX: Clean up attrset style

* SPDX: Include FSF-libre when determining free-ness

In Nixpkgs any license considered free or open source by both the FSF
and the OSI is considered free. When implementing the SPDX parsing I
must've missed the `isFsfLibre` property. This is now fixed.

Fixes #1263

* SPDX: Provide default for isFsfLibre

The `isFsfLibre` property isn't always included, as opposed to the
`isOsiApproved` property. To avoid errors we default it to `false` if
it's missing.

I opted for the simpler, more verbose way to write this.
`lic.isFsfLibre or isOsiApproved` would work as well and reads well but
it's not immediately clear that we're testing whether the field exists
and defaulting rather than doing a boolean OR.